Update recruitment test

+++ backend/src/main/java/com/project/hrm/module/recruitment/service/impl/ApplicationServiceImpl.java
@@ -209,8 +209,10 @@ public class ApplicationServiceImpl implements ApplicationService {
         Application app = applicationRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Not found application!"));
         fileService.deletePDF(app.getCvUrl());
-        cvReviewService.deleteReview(id);
-        interviewService.deleteInterview(id);
+        if(!app.getStatus().equals(ApplicationStatus.APPLIED)){
+            cvReviewService.deleteReview(id);
+            interviewService.deleteInterview(id);
+        }
         UUID candidateId = app.getCandidate().getId();
         applicationRepository.delete(app);
         boolean check = applicationRepository.existsByCandidate_Id(candidateId);
